# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "For the parsing of analytics flows.",
        "title": "Analytics-Parser API",
        "contact": {},
        "license": {
            "name": "Apache-2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.14"
    },
    "basePath": "/",
    "paths": {
        "/flow/getinputs/{id}": {
            "get": {
                "description": "Gets the inputs and configuration of the given flow",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flow"
                ],
                "summary": "Get flow inputs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Flow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/flows_api.Cell"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/flow/{id}": {
            "get": {
                "description": "Gets the parsed flow data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flow"
                ],
                "summary": "Get flow info parsed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Flow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Pipeline"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "flows_api.Cell": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.ConfigValue"
                    }
                },
                "cost": {
                    "type": "integer"
                },
                "deploymentType": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "inPorts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "operatorId": {
                    "type": "string"
                },
                "outPorts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "position": {
                    "$ref": "#/definitions/lib.CellPosition"
                },
                "source": {
                    "$ref": "#/definitions/lib.CellLink"
                },
                "target": {
                    "$ref": "#/definitions/lib.CellLink"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "lib.CellLink": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "magnet": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                }
            }
        },
        "lib.CellPosition": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
        "lib.ConfigValue": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "lib.DownstreamConfig": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                }
            }
        },
        "lib.InputTopic": {
            "type": "object",
            "properties": {
                "filterType": {
                    "type": "string"
                },
                "filterValue": {
                    "type": "string"
                },
                "mappings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.Mapping"
                    }
                },
                "topicName": {
                    "type": "string"
                }
            }
        },
        "lib.Mapping": {
            "type": "object",
            "properties": {
                "dest": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "lib.Operator": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "integer"
                },
                "deploymentType": {
                    "type": "string"
                },
                "downstream": {
                    "$ref": "#/definitions/lib.DownstreamConfig"
                },
                "id": {
                    "type": "string"
                },
                "imageId": {
                    "type": "string"
                },
                "inputTopics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.InputTopic"
                    }
                },
                "name": {
                    "type": "string"
                },
                "operatorId": {
                    "type": "string"
                },
                "upstream": {
                    "$ref": "#/definitions/lib.UpstreamConfig"
                }
            }
        },
        "lib.Pipeline": {
            "type": "object",
            "properties": {
                "flowId": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "operators": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/lib.Operator"
                    }
                }
            }
        },
        "lib.UpstreamConfig": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                }
            }
        }
    }
}