# lib/models.go
/*
 * Copyright 2025 InfAI (CC SES)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package lib

type Pipeline struct {
	FlowId    string              `json:"flowId,omitempty"`
	Image     string              `json:"image,omitempty"`
	Operators map[string]Operator `json:"operators,omitempty"`
}

type UpstreamConfig struct {
	Enabled bool
}

type DownstreamConfig struct {
	Enabled bool
}

type Operator struct {
	Id               string `json:"id,omitempty"`
	Name             string `json:"name,omitempty"`
	OperatorId       string `json:"operatorId,omitempty"`
	DeploymentType   string `json:"deploymentType,omitempty"`
	ImageId          string
	InputTopics      []InputTopic
	Cost             uint             `json:"cost"`
	UpstreamConfig   UpstreamConfig   `json:"upstream"`
	DownstreamConfig DownstreamConfig `json:"downstream"`
}

type InputTopic struct {
	TopicName   string
	FilterType  string
	FilterValue string
	Mappings    []Mapping
}

type Mapping struct {
	Source string
	Dest   string
}