# pkg/flows-api/entities.go
/*
 * Copyright 2018 InfAI (CC SES)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package flows_api

import (
	flowModels "github.com/SENERGY-Platform/analytics-flow-repo-v2/lib"
)

type Flow struct {
	flowModels.Flow
	Model Model `json:"model,omitempty"`
}

type Model struct {
	Cells []Cell `json:"cells,omitempty"`
}

type Cell flowModels.Cell